[FEAT/#10] 내역 추가 화면 개발 완료

+++ data/src/main/java/com/aos/data/entity/request/book/PostBooksLinesEntity.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PostBooksLinesEntity(
-    val money: Int,
+    val money: Double,
     val flow: String,
     val asset:String,
     val line: String,
